[EME][Mac] MediaKeys.createSession() fails with initData containing a contentId whose length is > 1/2 the initData. https://bugs.webkit.org/show_bug.cgi?id=153517 <rdar://problem/24303782>
Reviewed by Eric Carlson.

The length of contentId is given in bytes, not Uint16 characters. Use the former when extracting
the contentId string from the initData.

* platform/graphics/avfoundation/MediaPlayerPrivateAVFoundation.cpp:
(WebCore::MediaPlayerPrivateAVFoundation::extractKeyURIKeyIDAndCertificateFromInitData):


Canonical link: https://commits.webkit.org/171938@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@196083 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/graphics/avfoundation/MediaPlayerPrivateAVFoundation.cpp
@@ -1042,7 +1042,7 @@ bool MediaPlayerPrivateAVFoundation::extractKeyURIKeyIDAndCertificateFromInitDat
     if (!status || offset + keyIDLength > initData->length())
         return false;
 
-    RefPtr<Uint16Array> keyIDArray = Uint16Array::create(initDataBuffer, offset, keyIDLength);
+    RefPtr<Uint8Array> keyIDArray = Uint8Array::create(initDataBuffer, offset, keyIDLength);
     if (!keyIDArray)
         return false;
